Fixing bug where we would sometimes try to delete files that didn't exist.

+++ src/webpack-system-register.js
@@ -110,9 +110,13 @@ WebpackSystemRegister.prototype.apply = function(compiler) {
 		// Based on https://github.com/webpack/webpack/blob/ded70aef28af38d1deb2ac8ce1d4c7550779963f/lib/WebpackSystemRegister.js
 		compilation.plugin("optimize-chunk-assets", (chunks, callback) => {
 			externalModuleFiles.forEach(file => {
-				fs.unlink(file, err => {
-					if (err) {
-						throw err;
+				fs.exists(file, exists => {
+					if (exists) {
+						fs.unlink(file, err => {
+							if (err) {
+								throw err;
+							}
+						});
 					}
 				});
 			});
